added even listener to signup button to show signup form box

+++ backend_api/app/controllers/elabs_controller.rb
@@ -24,7 +24,7 @@ class ElabsController < ApplicationController
           if params[:question_id]
                question = Question.find(params[:question_id])
                elab = question.elabs.build(elab_params)
-               elab.layman_id = current_opp_layman.id # hard coded for now
+               elab.layman_id = current_layman.id # hard coded for now
                if elab.save
                     render json: ElabSerializer.new(elab).serializable_hash
                else
@@ -38,10 +38,12 @@ class ElabsController < ApplicationController
           if params[:question_id]
                question = Question.find(params[:question_id])
                elab = question.elabs.find(params[:id])
-               if elab.update(elab_params)
-                    render json: ElabSerializer.new(elab).serializable_hash
-               else
-                    render json: { errors: elab.errors }
+               if authenticate_elab(elab)
+                    if elab.update(elab_params)
+                         render json: ElabSerializer.new(elab).serializable_hash
+                    else
+                         render json: { errors: elab.errors }
+                    end
                end
           end
      end
@@ -50,10 +52,12 @@ class ElabsController < ApplicationController
           if params[:question_id]
                question = Question.find(params[:question_id])
                elab = question.elabs.find(params[:id])
-               if elab.destroy 
-                    render json: { message: "Elaboration deleted" }
-               else
-                    render json: { errors: elab.errors }
+               if authenticate_elab(elab)
+                    if elab.destroy 
+                         render json: { message: "Elaboration deleted" }
+                    else
+                         render json: { errors: elab.errors }
+                    end
                end
           end
      end
