+++ backend_api/app/controllers/replies_controller.rb
@@ -17,13 +17,13 @@ class RepliesController < ApplicationController
           if params[:elab_id]
                elab = Elab.find(params[:elab_id])
                reply = elab.replies.find(params[:id])
-               # if reply.layman_id == current_layman.id
+               if authenticate_reply(reply)
                     if reply.update(reply_params)
                          render json: ReplySerializer.new(reply).serializable_hash
                     else
                          render json: { message: "It seems like this post doesn't belong to you" }
                     end
-               # end
+               end
           end
      end
 
@@ -31,19 +31,19 @@ class RepliesController < ApplicationController
           if params[:elab_id]
                elab = Elab.find(params[:elab_id])
                reply = elab.replies.find(params[:id])
-               # if reply.layman == current_layman
+               if authenticate_reply(reply)
                     reply.destroy
                     render json: { message: "Reply deleted" }
-               # else
-               #      render json: { message: "It seems like this post doesn't belong to you" }
-               # end
+               else
+                    render json: { message: "It seems like this reply doesn't belong to you" }
+               end
           end
      end
 
      private 
 
      def reply_params
-          params.permit(:content, :elab_id, :layman_id)
+          params.permit(:content, :elab_id)
      end
 
 end
